add more explicit/clear exports
